- Move carmirror package to oldcarmirror - Create new carmirror package with bare min, in prep for pulling in go-car-mirror
