+++ oldcarmirror/payload_test.go
@@ -1,4 +1,4 @@
-package payload
+package oldcarmirror
 
 import (
 	"reflect"
