+++ oldcarmirror/session.go
@@ -1,4 +1,4 @@
-package carmirror
+package oldcarmirror
 
 import (
 	"context"
