+++ oldcarmirror/version.go
@@ -1,3 +1,3 @@
-package carmirror
+package oldcarmirror
 
 const Version = "0.1.0"
